Lobo Evolution 2.0

+++ src/main/java/org/loboevolution/laf/FontKey.java
@@ -1,3 +1,23 @@
+/*
+ * GNU GENERAL LICENSE
+ * Copyright (C) 2014 - 2021 Lobo Evolution
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * verion 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Contact info: ivan.difrancesco@yahoo.it
+ */
+
 package org.loboevolution.laf;
 
 import java.awt.Font;
@@ -9,8 +29,8 @@ import org.loboevolution.common.Objects;
 /**
  * The Class FontKey.
  *
- * @author utente
- * @version $Id: $Id
+ *
+ *
  */
 public class FontKey {
 
@@ -212,7 +232,7 @@ public class FontKey {
 			fs = "";
 		}
 		final Integer ss = this.superscript;
-		ch = ff.hashCode() ^ fw.hashCode() ^ fs.hashCode() ^ (int) this.fontSize ^ (ss == null ? 0 : ss.intValue());
+		ch = ff.hashCode() ^ fw.hashCode() ^ fs.hashCode() ^ (int) this.fontSize ^ (ss == null ? 0 : ss);
 		this.cachedHash = ch;
 		return ch;
 	}
